FTApi-M: (fix) 'abv' table size fix

+++ api/migrations/20230208214022-create-beer.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       abv: {
-        type: Sequelize.DOUBLE(18, 17),
+        type: Sequelize.DOUBLE,
         allowNull: false
       },
       address: {
